s_customer_multi_user: move domain maker to invader partner
