+++ shopinvader_customer_multi_user/models/shopinvader_partner.py
@@ -3,6 +3,7 @@
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import api, fields, models
+from odoo.osv import expression
 from odoo.tools.safe_eval import safe_eval
 
 
@@ -108,3 +109,39 @@ class ShopinvaderPartner(models.Model):
         ):
             partner.assign_invader_user_token()
         return binding
+
+    def _make_partner_domain(self, partner_field):
+        """Build partner domain for current shopinvader partner.
+
+        Domain is built considering backend's policy.
+        Main users can see all records but simple users
+        should be able to see records only from their parents
+        and not from sibling users.
+        """
+        # Main users can always see everything down their hierarchy
+        if self._is_admin_account() or self._is_main_account():
+            return [(partner_field, "child_of", self.record_id.id)]
+
+        # Change partner domain based on backend policy
+        policy_field = self.backend_id.multi_user_records_policy
+        related_partner = self[policy_field]
+        main_account = self.main_partner_id
+        # Normally see records owned by the same user
+        partner_domains = [[(partner_field, "=", self.record_id.id)]]
+        if related_partner and not self.record_id == related_partner:
+            # See records from its related_partner (normally the parent)
+            # but not other belonging to other users (like w/ child_of)
+            partner_domains.append([(partner_field, "=", related_partner.id)])
+            if main_account and main_account != related_partner:
+                # See records from parent main account
+                # NOTE: if main_account is the company,
+                # they see records from the company as well.
+                partner_domains.append([(partner_field, "=", main_account.id)])
+        return expression.OR(partner_domains)
+
+    def _is_admin_account(self):
+        # Root account
+        return not self.parent_id
+
+    def _is_main_account(self):
+        return self.record_id == self.main_partner_id
